rename: 배너 파일을 공통으로 쓰기 위해 폴더 이동 features/media/common => features/common
